requirements.txt
beautifulsoup4==4.10.0
matplotlib==3.5.2
mysql-connector-python==8.0.28
numpy==1.22.3
pandas==1.4.2
pyowm==2.10.0
requests==2.20.0
pygame==2.1.2
instabot==0.117.0
responses==0.21.0
webdriver_manager>=0.0
selenium==3.14.1
python-telegram-bot==20.0a1
dvb==1.2
rsa==4.8